Consolidate vector stores into single module
