+++ src/__init__.py
@@ -31,12 +31,16 @@ from .distributed_trainer import DistributedTrainer, MemoryConfig
 from .remote_memory import RemoteMemory
 from .memory_client_base import MemoryClientBase
 from .edge_memory_client import EdgeMemoryClient
-from .vector_store import VectorStore, FaissVectorStore
-from .encrypted_vector_store import EncryptedVectorStore
-from .pq_vector_store import PQVectorStore
-from .async_vector_store import AsyncFaissVectorStore
-from .ephemeral_vector_store import EphemeralVectorStore
-from .holographic_vector_store import HolographicVectorStore
+from .vector_stores import (
+    VectorStore,
+    FaissVectorStore,
+    LocalitySensitiveHashIndex,
+    EncryptedVectorStore,
+    EphemeralVectorStore,
+    HolographicVectorStore,
+    PQVectorStore,
+    AsyncFaissVectorStore,
+)
 from .dnc_memory import DNCMemory
 from .iter_align import IterativeAligner
 from .critic_rlhf import CriticScorer, CriticRLHFTrainer
