+++ tests/test_quantum_memory_server.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
 
-from asi.vector_store import VectorStore
+from asi.vector_stores import VectorStore
 from asi.quantum_memory_server import QuantumMemoryServer
 from asi.quantum_memory_client import QuantumMemoryClient
